Removed VisaManager component and refactored it into VisaRequirementMessage.jsx
