+++ pages/destinations/[destination].jsx
@@ -4,13 +4,14 @@ import Footer from '../../components/Footer.jsx'
 import PageHead from '../../components/PageHead.jsx'
 import EntryDetails from '../../components/destinations/EntryDetails.jsx'
 import SearchBar from '../../components/SearchBar.jsx'
+import VisaRequirementMessage from '../../components/destinations/VisaRequirementMessage.jsx'
 import VisaInfoCard from '../../components/destinations/VisaInfoCard.jsx'
 import TravelBan from '../../components/TravelBan.jsx'
 import Schengen from '../../components/Schengen.jsx'
 import KETA from '../../components/KETA.jsx'
 import IraqiKurdistanRegion from '../../components/IraqiKurdistanRegion.jsx'
 import ElectronicVisaApplyButton from '../../components/ElectronicVisaApplyButton.jsx'
-import { displayVisaMessage, displayTravelBan, displaySchengenInfo, displayKETA, displayIraqiKurdistan, displayEVisaButton } from '../../utils/VisaCardInfoManager.jsx'
+import { displayVisaMessage, displayTravelBan, displaySchengenInfo, displayKETA, displayIraqiKurdistan, displayEVisaButton } from '../../components/destinations/VisaRequirementMessage.jsx'
 import { countries, isoCodesList, schengenCountries } from '../../utils/countrydata.js'
 import { getVisaRequirements } from '../../utils/visadata.js'
 import { albula, supremeBold, supremeMedium } from '../../utils/localNextFonts.js'
@@ -52,15 +53,14 @@ export default function Destination({ requirement, passport, destination }){
                             <h1 className={"mt-20 text-4xl md:text-6xl md:mr-52"}>{destination}</h1>
                         </div>
                         <div className={"max-md:w-[96%] w-11/12 max-md:pl-3.5 md:mx-4 my-2"}>
-                            {displayVisaMessage(passport, destination, requirement)}
+                            <VisaRequirementMessage passport={passport} requirement={requirement} destination={destination}></VisaRequirementMessage>
                         </div>
                         <EntryDetails requirement={requirement}></EntryDetails>
                     </div>
 
                     <div className={supremeRegular.className + " grid md:grid-cols-2 md:grid-rows-4 gap-x-20 mr-3.5 md:mt-24 mb-20 md:mb-24 gap-x-56"}>
                         <span className={"flex flex-col gap-y-6 row-span-3 mt-8 md:mt-16"}>
-                            {requirement == "no admission" ? <TravelBan/> : <VisaInfoCard requirement={requirement}/>
-                            }
+                            {requirement == "no admission" ? <TravelBan/> : <VisaInfoCard requirement={requirement}/>}
                             <Schengen destination={destination}/>
                             <KETA destination={destination} requirement={requirement}></KETA>
                             <IraqiKurdistanRegion destination={destination} requirement={requirement}/>
